nop bai tap

+++ src/ss7/bai_tap/bai_tap1/bai2.js
@@ -1,4 +1,4 @@
 inputC= prompt("Nhập độ C: ");
-var C=parseInt(inputC);
+var C=parseFloat(inputC);
 var F=C*9/5 +32;
 alert("Độ F: " +F);
